pyproject.toml
[project]
name = "cbioportal-revamp"
version = "0.1.0"
description = "A modern revamp of cBioPortal using FastAPI, HTMX, and DuckDB."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn",
    "duckdb",
    "typer",
    "jinja2",
    "requests",
    "python-multipart",
    "redis",
    "python-dotenv",
    "psutil",
    "rich>=14.3.3",
    "tqdm>=4.67.3",
    "pyyaml>=6.0.3",
    "scipy",
    "httpx>=0.28",
    "questionary>=2.0",
    "prompt-toolkit>=3.0",
    "tomli-w>=1.0",
    "pydantic>=2.0",
]

[project.scripts]
cbio = "cbioportal.cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/cbioportal"]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pillow>=10.0",
    "playwright>=1.58.0",
    "pytest>=9.0.2",
]
